requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
langchain==0.0.325
torch==2.1.0
transformers==4.34.0
tiktoken==0.5.1
pytest==7.4.2
python-magic==0.4.27
tqdm==4.66.1
mistralai>=0.0.7
streamlit>=1.32.0 